s6canmsg lib test, fix missing C file, fix sysroot path

+++ components/misc/tools/files/lib/setsockopt.c
@@ -12,7 +12,7 @@
 
 extern ssize_t __real_setsockopt(int s, int level, int optname, const void *optval, socklen_t optlen);
 
-int __wrap_setsockopt(int s, int level, int optname, const void *optval, socklen_t optlen) {
+ssize_t __wrap_setsockopt(int s, int level, int optname, const void *optval, socklen_t optlen) {
     strerr_warni1x("wrap 'setsockopt' syscall");
     
     if(level!=SOL_CAN_RAW) {
